ai/requirements.txt
openai==1.12.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.0
numpy==1.26.3
